# docs/resultados/api_evidence/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThriveMind API",
    "description": "Ecosistema de bienestar holístico orquestado por IA",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/mente/generar": {
      "post": {
        "tags": [
          "Pilar Mente"
        ],
        "summary": "Generar sesión de meditación personalizada",
        "operationId": "generar_sesion_api_v1_mente_generar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeditacionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mente/historial": {
      "get": {
        "tags": [
          "Pilar Mente"
        ],
        "summary": "Historial de meditaciones",
        "operationId": "get_historial_api_v1_mente_historial_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/preferences/": {
      "get": {
        "tags": [
          "Preferencias"
        ],
        "summary": "Obtener configuración de pilares del usuario",
        "operationId": "get_preferences_api_v1_preferences__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Preferencias"
        ],
        "summary": "Actualizar configuración de pilares",
        "operationId": "update_preferences_api_v1_preferences__patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeditacionRequest": {
        "properties": {
          "intencion": {
            "type": "string",
            "title": "Intencion"
          },
          "objetivo": {
            "type": "string",
            "title": "Objetivo"
          },
          "duracion_min": {
            "type": "integer",
            "title": "Duracion Min",
            "default": 10
          },
          "checkin_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Id"
          },
          "generar_audio": {
            "type": "boolean",
            "title": "Generar Audio",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "intencion",
          "objetivo"
        ],
        "title": "MeditacionRequest"
      },
      "PillarConfig": {
        "properties": {
          "activo": {
            "type": "boolean",
            "title": "Activo",
            "default": true
          },
          "intensidad": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Intensidad",
            "description": "1=básico, 2=intermedio, 3=avanzado",
            "default": 1
          }
        },
        "type": "object",
        "title": "PillarConfig",
        "description": "Configuración de un pilar individual."
      },
...
      "UserPreferencesResponse": {
        "properties": {
          "mente_activo": {
            "type": "boolean",
            "title": "Mente Activo"
          },
          "mente_intensidad": {
            "type": "integer",
            "title": "Mente Intensidad"
          },
          "cuerpo_activo": {
            "type": "boolean",
            "title": "Cuerpo Activo"
          },
          "cuerpo_intensidad": {
            "type": "integer",
            "title": "Cuerpo Intensidad"
          },
          "entorno_activo": {
            "type": "boolean",
            "title": "Entorno Activo"
          },
          "entorno_intensidad": {
            "type": "integer",
            "title": "Entorno Intensidad"
          },
          "objetivo_principal": {
            "type": "string",
            "title": "Objetivo Principal"
          },
          "frecuencia_checkin": {
            "type": "string",
            "title": "Frecuencia Checkin"
          },
          "alergias": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alergias",
            "default": []
          },
          "preferencia_dieta": {
            "type": "string",
            "title": "Preferencia Dieta",
            "default": "sin_restriccion"
          },
          "presupuesto_semanal": {
            "type": "string",
            "title": "Presupuesto Semanal",
            "default": "medio"
          },
          "objetivo_fitness": {
            "type": "string",
            "title": "Objetivo Fitness",
            "default": "mantener"
          }
        },
        "type": "object",
        "required": [
          "mente_activo",
          "mente_intensidad",
          "cuerpo_activo",
          "cuerpo_intensidad",
          "entorno_activo",
          "entorno_intensidad",
          "objetivo_principal",
          "frecuencia_checkin"
        ],
        "title": "UserPreferencesResponse",
        "description": "Schema de respuesta con el estado completo de preferencias."
      },
      "UserPreferencesUpdate": {
        "properties": {
          "mente": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PillarConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "cuerpo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PillarConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "entorno": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PillarConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "objetivo_principal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "equilibrio",
                  "reducir_estres",
                  "mejorar_sueno",
                  "aumentar_energia",
                  "conexion_naturaleza",
                  "rendimiento_cognitivo"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Objetivo Principal"
          },
          "frecuencia_checkin": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "diario",
                  "cada_dos_dias",
                  "semanal"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Frecuencia Checkin"
          },
          "alergias": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alergias",
            "description": "Alergias alimentarias: gluten, lactosa, frutos_secos, mariscos, huevo, soja"
          },
          "preferencia_dieta": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "omnivora",
                  "vegetariana",
                  "vegana",
                  "pescatariana",
                  "sin_restriccion"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferencia Dieta"
          },
          "presupuesto_semanal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "bajo",
                  "medio",
                  "alto"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Presupuesto Semanal",
            "description": "bajo=<60€, medio=60-120€, alto=>120€"
          },
          "objetivo_fitness": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "mantener",
                  "perder_peso",
                  "ganar_musculo",
                  "tonificar",
                  "resistencia"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Objetivo Fitness"
          }
        },
        "type": "object",
        "title": "UserPreferencesUpdate",
        "description": "Schema para actualizar preferencias — todos los campos son opcionales."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}